+++ hedera-node/hedera-smart-contract-service-impl/src/main/java/com/hedera/node/app/service/contract/impl/exec/systemcontracts/hts/claimairdrops/TokenClaimAirdropDecoder.java
@@ -18,24 +18,19 @@ package com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.claim
 
 import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_TOKEN_ID;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.PENDING_AIRDROP_ID_LIST_TOO_LONG;
-import static com.hedera.node.app.service.contract.impl.utils.ConversionUtils.asTokenId;
+import static com.hedera.node.app.service.contract.impl.exec.utils.PendingAirdropsHelper.decodePendingAirdrops;
+import static com.hedera.node.app.service.contract.impl.exec.utils.PendingAirdropsHelper.pendingFTAirdrop;
+import static com.hedera.node.app.service.contract.impl.exec.utils.PendingAirdropsHelper.pendingNFTAirdrop;
 import static com.hedera.node.app.spi.workflows.HandleException.validateFalse;
 import static com.hedera.node.app.spi.workflows.HandleException.validateTrue;
 
 import com.esaulpaugh.headlong.abi.Address;
 import com.esaulpaugh.headlong.abi.Tuple;
-import com.hedera.hapi.node.base.AccountID;
-import com.hedera.hapi.node.base.NftID;
-import com.hedera.hapi.node.base.PendingAirdropId;
-import com.hedera.hapi.node.base.TokenID;
-import com.hedera.hapi.node.base.TokenType;
 import com.hedera.hapi.node.token.TokenClaimAirdropTransactionBody;
 import com.hedera.hapi.node.transaction.TransactionBody;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.HtsCallAttempt;
 import com.hedera.node.config.data.TokensConfig;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.ArrayList;
-import java.util.Arrays;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
@@ -60,28 +55,10 @@ public class TokenClaimAirdropDecoder {
         final var call = TokenClaimAirdropTranslator.CLAIM_AIRDROP.decodeCall(attempt.inputBytes());
         final var maxPendingAirdropsToClaim =
                 attempt.configuration().getConfigData(TokensConfig.class).maxAllowedPendingAirdropsToClaim();
-        validateFalse(((Tuple[]) call.get(0)).length > maxPendingAirdropsToClaim, PENDING_AIRDROP_ID_LIST_TOO_LONG);
-
         final var transferList = (Tuple[]) call.get(TRANSFER_LIST);
-        final var pendingAirdrops = new ArrayList<PendingAirdropId>();
-        Arrays.stream(transferList).forEach(transfer -> {
-            final var senderAddress = (Address) transfer.get(SENDER);
-            final var receiverAddress = (Address) transfer.get(RECEIVER);
-            final var tokenAddress = (Address) transfer.get(TOKEN);
-            final var serial = (long) transfer.get(SERIAL);
+        validateFalse(transferList.length > maxPendingAirdropsToClaim, PENDING_AIRDROP_ID_LIST_TOO_LONG);
 
-            final var senderId = attempt.addressIdConverter().convert(senderAddress);
-            final var receiverId = attempt.addressIdConverter().convert(receiverAddress);
-            final var tokenId = asTokenId(tokenAddress);
-
-            final var token = attempt.enhancement().nativeOperations().getToken(tokenId.tokenNum());
-            validateTrue(token != null, INVALID_TOKEN_ID);
-            if (token.tokenType().equals(TokenType.FUNGIBLE_COMMON)) {
-                pendingAirdrops.add(pendingFTAirdrop(senderId, receiverId, tokenId));
-            } else {
-                pendingAirdrops.add(pendingNFTAirdrop(senderId, receiverId, tokenId, serial));
-            }
-        });
+        final var pendingAirdrops = decodePendingAirdrops(attempt, transferList);
 
         return TransactionBody.newBuilder()
                 .tokenClaimAirdrop(TokenClaimAirdropTransactionBody.newBuilder().pendingAirdrops(pendingAirdrops))
@@ -120,27 +97,6 @@ public class TokenClaimAirdropDecoder {
         return TransactionBody.newBuilder()
                 .tokenClaimAirdrop(TokenClaimAirdropTransactionBody.newBuilder()
                         .pendingAirdrops(pendingNFTAirdrop(senderId, receiverId, token, serial)))
-                .build();
-    }
-
-    private PendingAirdropId pendingFTAirdrop(
-            @NonNull final AccountID senderId, @NonNull final AccountID receiverId, @NonNull final TokenID tokenId) {
-        return PendingAirdropId.newBuilder()
-                .senderId(senderId)
-                .receiverId(receiverId)
-                .fungibleTokenType(tokenId)
-                .build();
-    }
-
-    private PendingAirdropId pendingNFTAirdrop(
-            @NonNull final AccountID senderId,
-            @NonNull final AccountID receiverId,
-            @NonNull final TokenID tokenId,
-            final long serial) {
-        return PendingAirdropId.newBuilder()
-                .senderId(senderId)
-                .receiverId(receiverId)
-                .nonFungibleToken(NftID.newBuilder().tokenId(tokenId).serialNumber(serial))
                 .build();
     }
 }
